Broke tests up into individual files
